Correctly determine whether any requests are outstanding
Relying purely on whether the request promises have been fulfilled to determine
whether any requests were outstanding left a few edge cases:

 - When the request is fulfilled during the renderer's paused state
 - When the request fails
 - When the request is cancelled

In these cases, the fetcher would report that there are no outstanding requests
and the renderer would assume that any errors were genuine. At least in the
first case, this was incorrect.

This commit contains two fixes: first, to check if the request promises are NOT
PENDING rather than fulfilled. This treats errors and cancellations as NOT
OUTSTANDING. Second, to track when a request promise is resolved but not yet
read. This ensures that when the final request(s) are fulfilled while the
renderer is paused, the fetcher will correctly report that there are requests
the renderer has yet to read the result of, causing it to call Fetcher.request
again.